% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_func.R
\name{ggforest}
\alias{ggforest}
\title{Meta-analysis of multiple predictors in a list of models with the same structure.}
\usage{
ggforest(x, intercept = F, labels = NULL, hetero = "none",
  palette = wesanderson::wes_palettes$Darjeeling1)
}
\arguments{
\item{x}{The list returned by the meta_models function.}

\item{intercept}{A Boolean. Do you want to plot the model intercept? Defaults to \code{FALSE}.}

\item{labels}{A vector same length as number of predictors to contain pretty names for the predictors.}

\item{hetero}{A character string with the desired heterogeneity measure to be displayed.
Options: "none" {default}; "I2"; "Q"}

\item{palette}{A vector containing hex codes for the desired colour palette. Defaults to Darjeeling1 from
\code{\link{wes_palettes}}.}
}
\value{
Returns a dataframe for quick reference and a list of the full
 meta-analysis outputs from \code{\link{metafor}}.
}
\description{
\code{ggforest} returns a list containing metafor outputs for each predictor and a talbe of the combined results.
}
\details{
This function takes the list returned by .
}
