% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_func.R
\name{meta_models}
\alias{meta_models}
\title{Meta-analysis of multiple predictors in a list of models with the same structure.}
\usage{
meta_models(model_list)
}
\arguments{
\item{model_list}{A list of lmer or glmer models with the same structure.}
}
\value{
Returns a dataframe for quick reference and a list of the full
 meta-analysis outputs from \code{\link{metafor}}.
}
\description{
\code{meta_models} returns a list containing metafor outputs for each predictor and a table of the combined results.
}
\details{
This code currently pulls the beta and standard error from the each predictor
in the model and uses the \code{\link{rma.uni()}} function from
\code{\link{metafor}} to run meta-analyses over each predictor.
}
